--- packages/utilities/web-vitals/__mocks__/react-adaptive-hooks/hardware-concurrency/index.js
@@ -1,6 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-const useHardwareConcurrency = () => {
-  return { numberOfLogicalProcessors: 4 };
-};
-
-export { useHardwareConcurrency };
