+++ packages/utilities/web-vitals/__mocks__/react-adaptive-hooks/index.js
@@ -1,3 +1,13 @@
-export { useNetworkStatus } from './network';
-export { useHardwareConcurrency } from './hardware-concurrency';
-export { useMemoryStatus } from './memory';
+const useNetworkStatus = () => {
+  return { effectiveConnectionType: '4g' };
+};
+
+const useMemoryStatus = () => {
+  return { deviceMemory: 3 };
+};
+
+const useHardwareConcurrency = () => {
+  return { numberOfLogicalProcessors: 4 };
+};
+
+export { useNetworkStatus, useHardwareConcurrency, useMemoryStatus };
