--- packages/utilities/web-vitals/__mocks__/react-adaptive-hooks/memory/index.js
@@ -1,6 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-const useMemoryStatus = () => {
-  return { deviceMemory: 3 };
-};
-
-export { useMemoryStatus };
