--- packages/utilities/web-vitals/__mocks__/react-adaptive-hooks/network/index.js
@@ -1,6 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-const useNetworkStatus = () => {
-  return { effectiveConnectionType: '4g' };
-};
-
-export { useNetworkStatus };
