--- packages/utilities/web-vitals/babel.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  presets: ['@babel/preset-env'],
-};
